+++ packages/assignment-5/src/refactoring/components/product/ProductContainer.tsx
@@ -17,22 +17,29 @@ export const ProductContainer = ({ products }: ProductContainerProps) => {
   };
 
   return (
-    <>
-      {products.map((product) => {
-        const remainingStock = getRemainingStock(cart, product);
-        return (
-          <ProductCard
-            key={product.id}
-            product={product}
-            remainingStock={remainingStock}
-          >
-            <AddProductButton
-              onClick={() => handleAddProductButtonClick(product)}
-              {...useAddProductButtonProps(remainingStock)}
-            />
-          </ProductCard>
-        );
-      })}
-    </>
+    <div>
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+        <div>
+          <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
+          <div className="space-y-2">
+            {products.map((product) => {
+              const remainingStock = getRemainingStock(cart, product);
+              return (
+                <ProductCard
+                  key={product.id}
+                  product={product}
+                  remainingStock={remainingStock}
+                >
+                  <AddProductButton
+                    onClick={() => handleAddProductButtonClick(product)}
+                    {...useAddProductButtonProps(remainingStock)}
+                  />
+                </ProductCard>
+              );
+            })}
+          </div>
+        </div>
+      </div>
+    </div>
   );
 };
